+++ retdec/tools/decompiler.py
@@ -295,6 +295,13 @@ def parse_args(argv):
         help='should the file be compiled with debugging information?'
     )
     parser.add_argument(
+        '-s', '--compiler-strip',
+        dest='comp_strip',
+        action='store_true',
+        default=None,
+        help='should the file be stripped after compilation?'
+    )
+    parser.add_argument(
         '-f', '--file-format',
         dest='file_format',
         choices=['elf', 'pe'],
@@ -406,6 +413,7 @@ def main(argv=None):
     add_decompilation_param_when_given(args, params, 'comp_compiler')
     add_decompilation_param_when_given(args, params, 'comp_optimizations')
     add_decompilation_param_when_given(args, params, 'comp_debug')
+    add_decompilation_param_when_given(args, params, 'comp_strip')
     add_decompilation_param_when_given(args, params, 'generate_archive')
     decompilation = decompiler.start_decompilation(**params)