+++ tests/decompiler_tests.py
@@ -178,6 +178,15 @@ class DecompilerStartDecompilationTests(BaseServiceTests):
             params=AnyParamsWith(comp_debug=True)
         )
 
+    def test_comp_strip_is_set_to_correct_value_when_given(self):
+        self.start_decompilation_with_any_input_file(
+            comp_strip=True
+        )
+
+        self.assert_post_request_was_sent_with(
+            params=AnyParamsWith(comp_strip=True)
+        )
+
     def test_adds_leading_dash_to_comp_optimizations_when_missing(self):
         self.start_decompilation_with_any_input_file(
             comp_optimizations='O1'