+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ dev
   source files.
 * Added support for selecting whether C files should be compiled with debugging
   information.
+* Added support for selecting whether compiled C files should be stripped.
 * Added support for printing script versions via the ``-V``/``--version``
   parameter.